requirements.txt
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tf2onnx==1.13.0
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.2.1
tokenizers>=0.19,<0.20
torch==2.4.0
torchmetrics==0.10.3
torchvision==0.14.1
tqdm==4.64.1
retina-face==0.0.13
scikit-image==0.19.3
scikit-learn==1.2.1
matplotlib==3.6.3
transformers
pytorch_lightning
albumentations
iglovikov_helper_functions
timm
asone-ocr
easy-ocr
gradio
gdown
Pillow==9.5.0
